LayoutTests/fast/css/text-overflow-ellipsis-hit-test.html

<!DOCTYPE html>
<html>
<head>
<style>
#test {
text-overflow: ellipsis;
width: 3em;
overflow: hidden;
white-space: nowrap;
position: absolute;
left: 10px;
top: 10px;
font-size: 20px;
font-family: ahem;
}
</style>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div id="test"><a href="#" onclick="this.textContent='clicked'">unclicked</a></div>
<script>
description('This tests the hit-testing of text-overflow:ellipsis. Click and hover events should be unaffected by the ellipsis.');

var test = document.getElementById("test");

if (window.eventSender) {
  eventSender.mouseMoveTo(40, 20);
  eventSender.mouseDown();
  eventSender.mouseUp();
}

shouldBeEqualToString("test.firstElementChild.textContent", "clicked");

// Clean up after ourselves
test.parentNode.removeChild(test);
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
