LayoutTests/fast/css/text-overflow-ellipsis-hit-test-expected.txt

This tests the hit-testing of text-overflow:ellipsis. Click and hover events should be unaffected by the ellipsis.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test.firstElementChild.textContent is "clicked"
PASS successfullyParsed is true

TEST COMPLETE
